+++ CatAlgo/src/main/java/grammarModel/impl/SyntacticChains.java
@@ -6,9 +6,9 @@ import grammarModel.ISyntacticChains;
 
 public class SyntacticChains extends Chains implements ISyntacticChains {
 
-	private List<Long> leafIDs;
+	private List<Integer> leafIDs;
 	
-	public SyntacticChains(List<List<String>> listOfChains, List<Long> leafIDs) {
+	public SyntacticChains(List<List<String>> listOfChains, List<Integer> leafIDs) {
 		super(listOfChains);
 		this.leafIDs = leafIDs;
 	}
